fix: enabled real-time ingestion and simplified pipeline
